requirements.txt
Flask==2.3.3
groq-sdk==0.10.0
python-dotenv==1.0.0
